refactor: kebab case for component file names
